refactor: move deposit and transfer to team account
